+++ NEWS.md
@@ -1,3 +1,7 @@
+# ggsvg 0.1.2 2021-12-31
+
+* Adjust fastpath such that grob names are always unique
+
 # ggsvg 0.1.1 2021-12-24 
 
 * Add fastpath when SVG is static i.e. reduce number of parsing attempts